corrections and update
