+++ test.cabal
@@ -44,8 +44,8 @@ library
                 , bytestring
   default-language: Haskell2010
 
-executable test-conllu
-  main-is: ConlluProcess.hs
+executable test-NLUJson
+  main-is: NLUJson.hs
   other-modules:
       Main1
       Main2
@@ -53,7 +53,7 @@ executable test-conllu
       Paths_test
   hs-source-dirs:
       app
-  ghc-options: -threaded -rtsopts -with-rtsopts=-N -main-is ConlluProcess
+  ghc-options: -threaded -rtsopts -with-rtsopts=-N -main-is NLUJson
   build-depends:
                     aeson
                 , base >=4.7 && <5
